requirements.txt
httpx
python-dotenv
fastapi
uvicorn
moviepy==1.0.3
pillow==9.5.0
requests
playwright
playwright-stealth
rembg
onnxruntime
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydub
elevenlabs
